% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TangledFeatures.R
\name{DataCleaning}
\alias{DataCleaning}
\title{Compute a correlation matrix for the given dataset}
\usage{
DataCleaning(Data)
}
\arguments{
\item{Data}{A data frame}
}
\description{
Compute a correlation matrix for the given dataset
}
\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
  # Retrieve group classification from hierarchical cluster analysis
  clustering <- cluster_analysis(iris[, 1:4], n = 3)

  # Goodness of group classification
  cluster_discrimination(clustering)
}

}
