% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TangledFeatures.R
\name{Housing_Prices_dataset}
\alias{Housing_Prices_dataset}
\alias{DataCleaning}
\title{Tree Distances}
\usage{
DataCleaning(Data)
}
\arguments{
\item{Data}{The imported Data Frame}
}
\value{
The cleaned data. Please note that the data has been cleaned including column names, NA values and coerced to data table
}
\description{
These datasets contain the distances between sets
of 10-tip, 11-tip and 12-tip trees.
}
\keyword{Automatic}
\keyword{Cleaning}
\keyword{Data}
\keyword{datasets}
