% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TangledFeatures.R
\name{GeneralCor}
\alias{GeneralCor}
\title{Generalized Correlation function}
\usage{
GeneralCor(df, cor1 = "pearson", cor2 = "PointBiserial", cor3 = "kendall")
}
\arguments{
\item{df}{The imported Data Frame}

\item{cor1}{The correlation metric between two continuous features. Defaults to pearson correlation}

\item{cor2}{The correlation metric between one categorical feature and one continuous feature. Defaults to biserial correlation correlation}

\item{cor3}{The correlation metric between two categorical features. Defaults to Cramers-V correlation}
}
\value{
Returns a correlation matrix containing the correlation values between the features
}
\description{
Generalized Correlation function
}
