% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TangledFeatures.R
\name{Housing_Prices_dataset}
\alias{Housing_Prices_dataset}
\alias{DataCleaning}
\title{Housing prices dataset}
\usage{
DataCleaning(Data, Y_var)
}
\arguments{
\item{Data}{The imported Data Frame}

\item{Y_var}{The X variable}
}
\value{
The cleaned data. Please note that the data has been cleaned including column names, NA values and coerced to data table
}
\description{
Housing prices dataset
}
\keyword{Automatic}
\keyword{Cleaning}
\keyword{Data}
\keyword{datasets}
